COSBETA-81: Move Sentry-Raven dependency and config to shared web
