+++ shared-web/shared-web.gemspec
@@ -15,6 +15,7 @@ Gem::Specification.new do |s|
   s.add_dependency "mini_magick", "4.9.3"
   s.add_dependency "rails", "~> 5.2"
   s.add_dependency "request_store", "1.4.1"
+  s.add_dependency "sentry-raven", "2.9.0"
   s.add_dependency "sidekiq", "5.2.5"
   s.add_dependency "sidekiq-cron", "1.1.0"
   s.add_dependency "slowpoke", "0.2.1"
